working with non-subclass
