+++ .gitignore
@@ -1,3 +1,4 @@
 venv
 Noto-unhinted
 pbt.egg-info
+__pycache__
